Add new Contact class. Update Address, Email, and Phone class to extend Contact.
